@@ -5,10 +5,17 @@
   "license": "MIT",
   "version": "0.3.4",
   "main": "./index.js",
+  "scripts": {
+    "size": "webpack --json | webpack-bundle-size-analyzer",
+    "build": "webpack",
+    "watch": "webpack --watch",
+    "server": "webpack-dev-server",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
   "babel": {
     "presets": [
       [
-        "env",
+        "@babel/env",
         {
           "targets": {
             "browsers": [
@@ -19,57 +26,51 @@
         }
       ],
       [
-        "react"
+        "@babel/react"
       ]
     ]
   },
   "dependencies": {
     "@beyonk/google-fonts-webpack-plugin": "^1.2.1",
     "@fortawesome/fontawesome-svg-core": "^1.2.14",
-    "@fortawesome/free-solid-svg-icons": "^5.6.3",
+    "@fortawesome/free-solid-svg-icons": "^5.7.1",
     "@fortawesome/react-fontawesome": "^0.1.4",
-    "bootstrap": "^3.3.7",
-    "bootswatch": "^3.3.7",
+    "babel-loader": "^8.0.5",
     "browser-request": "^0.3.3",
     "css-loader": "^2.1.0",
     "deep-equal": "^1.0.1",
     "deepcolour": "^2.0.0",
-    "fetch": "^1.1.0",
-    "highlight.js": "^9.5.0",
     "horten": "^0.9.0",
     "horten-websocket": "^0.2.8",
-    "inherits": "^2.0.1",
-    "js-yaml": "^3.10.0",
+    "js-yaml": "^3.12.1",
     "less": "^3.9.0",
-    "less-loader": "^4.0.3",
+    "less-loader": "^4.1.0",
     "lodash": "^4.17.4",
     "mini-css-extract-plugin": "^0.5.0",
-    "react": "^16.2.0",
-    "react-bootstrap": "^0.32.1",
-    "react-dom": "^16.2.0",
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
     "react-markdown": "^4.0.6",
     "react-syntax-highlighter": "^10.1.2",
     "react-textarea-autosize": "^7.1.0",
-    "request": "^2.82.0",
-    "request-promise": "^4.2.2",
     "string2png": "^0.5.3",
     "style-loader": "^0.23.1",
     "url-loader": "1.1.2"
   },
   "devDependencies": {
-    "babel-loader": "7",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
+    "@babel/core": "^7.0.0",
+    "@babel/preset-env": "^7.3.1",
+    "@babel/preset-react": "^7.0.0",
+    "babel-plugin-lodash": "^3.3.4",
+    "babel-preset-env": "^1.7.0",
     "babelify": "^10.0.0",
-    "browserify": "^16.2.3",
     "chai": "^4.2.0",
-    "copy-webpack-plugin": "^4.6.0",
-    "del": "^3.0.0",
+    "copy-webpack-plugin": "4.6.0",
     "eslint": "^5.13.0",
     "eslint-plugin-react": "^7.12.4",
-    "file-loader": "^3.0.1",
+    "file-loader": "3.0.1",
     "html-webpack-plugin": "^3.2.0",
     "webpack": "^4.29.1",
+    "webpack-bundle-size-analyzer": "^3.0.0",
     "webpack-cli": "^3.2.3",
     "webpack-dev-server": "^3.1.14"
   },
@@ -84,9 +85,7 @@
   "directories": {
     "test": "test"
   },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
+
   "repository": {
     "type": "git",
     "url": "git+https://github.com/koopero/horten-control.git"
