@@ -1,11 +1,13 @@
-const React = require('react')
+import React from 'react'
 import SyntaxHighlighter from 'react-syntax-highlighter'
 import { atelierLakesideDark, atelierLakesideLight } from 'react-syntax-highlighter/dist/styles/hljs'
 
-const jsYaml = require('js-yaml')
+import jsYaml from 'js-yaml'
 const safeDump = ( v ) => 'undefined' == typeof v ? '# undefined' : jsYaml.safeDump( v )
-module.exports = ( props ) => {
+const YAML = ( props ) => {
   return (
     <SyntaxHighlighter className="yaml" language={'yaml'} style={ props.highkey ? atelierLakesideLight : atelierLakesideDark }>{ safeDump ( props.data ) }</SyntaxHighlighter>
   )
 }
+
+export default YAML
