@@ -1,12 +1,11 @@
-'use strict'
-
-const Base = require('../Base')
-  , React = require('react')
+import Base from '../Base'
+import React from 'react'
 
 import Markdown from 'react-markdown'
 import { safeLoad } from 'js-yaml'
 import SyntaxHighlighter from 'react-syntax-highlighter'
 import { ocean } from 'react-syntax-highlighter/dist/styles/hljs'
+import Control from '../Control'
 
 const mdutil = require('./util')
 
@@ -15,20 +14,23 @@ require('./index.less')
 
 const CodeBlock = function ( props ) {
   const source = props.value || ''
-    , language = props.language
-    , meta = ( this && this.props && this.props.meta ) || {}
+  const language = props.language
+
+
+  const meta = ( this && this.props && this.props.meta ) || {}
 
   switch( language ) {
   case 'control':
     var config = safeLoad( source )
     config.meta = H.util.compose( meta, config.meta )
-    var Control = require('../Control')
+    // var Control = require('../Control').default
     return (
       <div className='md-element'>
         <Control {...config} />
       </div>
     )
     break
+
   default:
     return (
       <div className='md-element'>
@@ -51,7 +53,7 @@ const Heading = function( props ) {
   return <pre>{ JSON.stringify( props, null, 2 ) }</pre>
 }
 
-class OurMarkdown extends Base {
+export default class OurMarkdown extends Base {
   constructor( props ) {
     super( props )
     this.state.type = 'markdown'
@@ -66,7 +68,7 @@ class OurMarkdown extends Base {
       }
     )
 
-    console.log( '**********', Markdown.renderers )
+    // console.log( '**********', Markdown.renderers )
   }
 
 
@@ -89,5 +91,3 @@ class OurMarkdown extends Base {
     )
   }
 }
-
-module.exports = OurMarkdown
