@@ -1,9 +1,9 @@
-const util = exports
+const util = module.exports
 
-const _ = require('lodash')
-const urllib = require('url')
+import _ from 'lodash'
+import urllib from 'url'
 
-util.unique = function ( state, text ) {
+export function unique( state, text ) {
   let index = 0
   let result = text
   while( result in state ) {
@@ -15,13 +15,15 @@ util.unique = function ( state, text ) {
   return result
 } 
 
-util.hashTitle = ( text ) => text
-  .toLowerCase()
-  .trim()
-  .replace(/[\u2000-\u206F\u2E00-\u2E7F\\'!"#$%&()*+,./:;<=>?@[\]^`{|}~]/g, '')
-  .replace(/\s/g, '-') 
+export function hashTitle( text ) {
+  return String( text || '' )
+    .toLowerCase()
+    .trim()
+    .replace(/[\u2000-\u206F\u2E00-\u2E7F\\'!"#$%&()*+,./:;<=>?@[\]^`{|}~]/g, '')
+    .replace(/\s/g, '-' )
+} 
 
-util.makeIndex = function ( {
+export function makeIndex ( {
   markdown, idPrefix = '', path
 } ) {
   let index = []
@@ -31,11 +33,11 @@ util.makeIndex = function ( {
 
   let level = path.length
   let order = 0
-  let unique = util.unique.bind( null, {} ) 
+  let uniq = unique.bind( null, {} ) 
 
   markdown.split(/[\r\n]+/).map( line => line.replace( /^(\#+)\s*(.*)/g, ( match, tag, text ) => {
     let tagLevel = tag.length + level
-    let hash = unique( idPrefix + util.hashTitle( text ) )
+    let hash = uniq( idPrefix + hashTitle( text ) )
     path = path.slice( 0, tagLevel )
     path[tagLevel-1] = hash
     index.push( {
