@@ -1,14 +1,10 @@
-'use strict'
-module.exports = HortenControl
 require('./index.less')
 
+import _ from 'lodash'
+import React from 'react'
+import H from 'horten'
 
-const _ = require('lodash')
-  , React = require('react')
-  , H = require('horten')
-
-
-function HortenControl( props ) {
+export default function HortenControl( props ) {
   props = _.clone( props )
   const meta = new H.Mutant( props.meta )
 
@@ -24,19 +20,20 @@ function HortenControl( props ) {
 
   type = type || 'group'
 
-
   props.type = type
+
   do {
     var lastType = props.type
     let metaType = meta.get('type', props.type )
     _.defaults( props, metaType )
     _.merge( props, _.pick( metaType, ['type'] ) )
-
   } while( props.type != lastType )
 
   props.meta = meta.get()
 
   let _class = props._class
+
+  console.log( 'here?', _class )
 
   if ( !_class ) {
     return (
